No longer creating empty protoactivities
if the user opens the New Protoactivity screen and then just goes back

+++ ActRec/ActRec/Source/View/ProtoActivity_Editing_Layout.cs
@@ -8,6 +8,7 @@ using Xamarin.Forms;
 
 namespace ActivityRecommendation.View
 {
+    // Allows creating or updating a ProtoActivity
     public class ProtoActivity_Editing_Layout : ContainerLayout, OnBack_Listener
     {
         public ProtoActivity_Editing_Layout(ProtoActivity protoActivity, ProtoActivity_Database database)
@@ -18,12 +19,12 @@ namespace ActivityRecommendation.View
             this.textBox.Text = protoActivity.Text;
 
             Vertical_GridLayout_Builder gridBuilder = new Vertical_GridLayout_Builder();
-            gridBuilder.AddLayout(new TextblockLayout("ProtoActivity #" + protoActivity.Id));
-
-            Button deleteButton = new Button();
-            deleteButton.Text = "Delete";
-            deleteButton.Clicked += DeleteButton_Clicked;
-            gridBuilder.AddLayout(new ButtonLayout(deleteButton));
+            TextblockLayout title;
+            if (protoActivity.Id >= 0)
+                title = new TextblockLayout("Protoactivity #" + protoActivity.Id);
+            else
+                title = new TextblockLayout("New Protoactivity");
+            gridBuilder.AddLayout(title);
             gridBuilder.AddLayout(ScrollLayout.New(new TextboxLayout(this.textBox)));
 
             this.SubLayout = gridBuilder.Build();
@@ -31,19 +32,28 @@ namespace ActivityRecommendation.View
 
         private void DeleteButton_Clicked(object sender, EventArgs e)
         {
-            this.database.Remove(this.protoActivity);
-            this.protoActivity.Ratings = Distribution.MakeDistribution(0, 0, 0);
-            this.protoActivity.Text = null;
+            this.delete();
+        }
+        private void delete()
+        {
+            this.textBox.Text = null;
         }
 
         public void OnBack(LayoutChoice_Set layout)
         {
-            this.save();
+            this.persist();
         }
-        private void save()
+        private void persist()
         {
-            this.protoActivity.Text = this.textBox.Text;
-            this.protoActivity.LastInteractedWith = DateTime.Now;
+            if (this.protoActivity.Text != this.textBox.Text)
+            {
+                this.protoActivity.Text = this.textBox.Text;
+                this.protoActivity.LastInteractedWith = DateTime.Now;
+                if (this.protoActivity.Text != null && this.protoActivity.Text != "")
+                    this.database.Put(this.protoActivity);
+                else
+                    this.database.Remove(this.protoActivity);
+            }
         }
         private ProtoActivity protoActivity;
         private ProtoActivity_Database database;
